requirements.txt
fastapi~=0.116.1
uvicorn==0.35.0
SQLAlchemy>=2.0.30,<2.1
pydantic==2.11.1
pydantic-settings==2.10.1

#python-jose[cryptography]==3.5.0
PyJWT[crypto]==2.10.1

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

PyMySQL==1.1.1
alembic>=1.15.0,<1.17
python-dotenv

pytest==8.2.1
pytest-asyncio==0.23.6
httpx==0.27.0

geopandas==1.1.1
shapely==2.1.0